Refactored the code.

+++ client/Angular-Graduation-Project/src/app/authentication/components/register/register.component.ts
@@ -7,6 +7,9 @@ import { AuthService } from 'src/app/core/services/auth.service';
 import { Router } from '@angular/router';
 import { COMMA, ENTER } from '@angular/cdk/keycodes';
 import { passwordMatchValidator } from '../../services/password-match-validator.service';
+import { IJwt } from 'src/app/shared/interfaces/ijwt';
+import { IUser } from 'src/app/shared/interfaces/iuser';
+import { ITopic } from 'src/app/shared/interfaces/itopic';
 
 @Component({
   selector: 'app-register',
@@ -25,7 +28,7 @@ export class RegisterComponent implements OnInit {
   // default type of the form
   registerOrEdit = 'register';
   registerFormGroup: FormGroup;
-  user!: any;
+  user!: IUser | undefined;
 
   // init the form
   constructor(
@@ -86,29 +89,34 @@ export class RegisterComponent implements OnInit {
 
   ngOnInit(): void {
     this.apiCalls.getAllTopics().subscribe({
-      next: (data) => {
-        this.options = data.topics.map((dataObj: any) => dataObj.name);
+      next: (data: { topics: ITopic[] }) => {
+        this.options = data.topics.map((dataObj: ITopic) => dataObj.name);
       },
       error: (err) => console.error(err),
     });
 
     if (history.state && history.state.user) {
+      console.log(history.state.user);
+
       this.user = history.state.user;
       const topics = history.state.topics;
       this.registerOrEdit = 'edit';
-      this.registerFormGroup.patchValue({
-        personalDetailsGroup: {
-          name: this.user.name,
-          email: this.user.email,
-          description: this.user.description,
-        },
-        // prepopulate the form value for topics
-        topicsGroup: {
-          topics: topics.slice(),
-        },
-      });
-      // prepopulate the topics chips
-      this.topics = topics.slice();
+
+      if (this.user) {
+        this.registerFormGroup.patchValue({
+          personalDetailsGroup: {
+            name: this.user.name,
+            email: this.user.email,
+            description: this.user.description,
+          },
+          // prepopulate the form value for topics
+          topicsGroup: {
+            topics: topics.slice(),
+          },
+        });
+        // prepopulate the topics chips
+        this.topics = topics.slice();
+      }
     }
   }
 
@@ -131,7 +139,7 @@ export class RegisterComponent implements OnInit {
         console.log(sendData);
 
         this.apiCalls.postRegisterForm(sendData).subscribe({
-          next: (response: any) => {
+          next: (response: IJwt) => {
             this.authService.createSession(response);
             this.router.navigate(['/']);
           },
@@ -153,30 +161,32 @@ export class RegisterComponent implements OnInit {
           },
         });
       } else {
-        this.apiCalls.postEditUserForm(sendData, this.user._id).subscribe({
-          next: (response) => {
-            this.authService.destroySession();
-            alert('Your credentials have changed. Please login again.');
-            this.router.navigate(['/']);
-          },
-          error: (err) => {
-            console.error(err);
-            // we can still get that kind of mistake, even when editting. we can't set our name / email to taken vals;
-            // convenient way to set errors
-            if (err.status === 409) {
-              this.registerFormGroup
-                .get('personalDetailsGroup.name')
-                ?.setErrors({
-                  usernameOrEmailTaken: true,
-                });
-              this.registerFormGroup
-                .get('personalDetailsGroup.email')
-                ?.setErrors({
-                  usernameOrEmailTaken: true,
-                });
-            }
-          },
-        });
+        if (this.user) {
+          this.apiCalls.postEditUserForm(sendData, this.user._id).subscribe({
+            next: (response) => {
+              this.authService.destroySession();
+              alert('Your credentials have changed. Please login again.');
+              this.router.navigate(['/']);
+            },
+            error: (err) => {
+              console.error(err);
+              // we can still get that kind of mistake, even when editting. we can't set our name / email to taken vals;
+              // convenient way to set errors
+              if (err.status === 409) {
+                this.registerFormGroup
+                  .get('personalDetailsGroup.name')
+                  ?.setErrors({
+                    usernameOrEmailTaken: true,
+                  });
+                this.registerFormGroup
+                  .get('personalDetailsGroup.email')
+                  ?.setErrors({
+                    usernameOrEmailTaken: true,
+                  });
+              }
+            },
+          });
+        }
       }
     } else {
       console.error('Form has errors.');
